Create configurable couchdb admin, set configuration in couchdb, proxy couchdb to /db/ if enabled, and add make sure to forward http traffic to https when enabled

+++ develop.sh
@@ -44,12 +44,47 @@ docker build -t tangerine/tangerine:local .
 [ "$(docker ps -a | grep $T_CONTAINER_NAME)" ] && docker rm $T_CONTAINER_NAME 
 
 COUCHDB_OPTIONS=""
-if [ "$T_COUCHDB_ENABLE" = "true"  ]; then
-  docker stop $T_COUCHDB_CONTAINER_NAME
-  docker rm $T_COUCHDB_CONTAINER_NAME
-  docker run -d -v $(pwd)/data/couchdb:/opt/couchdb/data --name $T_COUCHDB_CONTAINER_NAME couchdb
+if [ "$T_COUCHDB_ENABLE" = "true" ] && [ "$T_COUCHDB_ENDPOINT" = "http://couchdb:5984" ]; then
+  if [ ! -d data/couchdb ]; then
+    mkdir data/couchdb
+  fi
+  if [ ! -d data/couchdb/data ]; then
+    mkdir data/couchdb/data
+  fi
+  if [ ! -d data/couchdb/local.d ]; then
+    mkdir data/couchdb/local.d
+  fi
+  if [ ! -f data/couchdb/local.d/local.ini ]; then
+    echo "
+[chttpd]
+bind_address = any
+
+[httpd]
+bind_address = any
+
+[couch_httpd_auth]
+require_valid_user = true
+
+[chttpd]
+require_valid_user = true
+    " > data/couchdb/local.d/local.ini
+  fi
+  [ "$(docker ps | grep $T_COUCHDB_CONTAINER_NAME)" ] && docker stop $T_COUCHDB_CONTAINER_NAME
+  [ "$(docker ps -a | grep $T_COUCHDB_CONTAINER_NAME)" ] && docker rm $T_COUCHDB_CONTAINER_NAME
+  docker run -d \
+     -e COUCHDB_USER=$T_COUCHDB_USER_ADMIN_NAME \
+     -e COUCHDB_PASSWORD=$T_COUCHDB_USER_ADMIN_PASS \
+     -p 5984:5984 \
+     -v $(pwd)/data/couchdb/data:/opt/couchdb/data \
+     -v $(pwd)/data/couchdb/local.d:/opt/couchdb/etc/local.d \
+     --name $T_COUCHDB_CONTAINER_NAME \
+     couchdb
   COUCHDB_OPTIONS="
     --link $T_COUCHDB_CONTAINER_NAME:couchdb \
+    -e T_COUCHDB_ENABLE=$T_COUCHDB_ENABLE \
+    -e T_COUCHDB_ENDPOINT=$T_COUCHDB_ENDPOINT \
+    -e T_COUCHDB_USER_ADMIN_NAME=$T_COUCHDB_USER_ADMIN_NAME \
+    -e T_COUCHDB_USER_ADMIN_PASS=$T_COUCHDB_USER_ADMIN_PASS \
   "
 fi
